+++ website/templates/call_categories/call_categories.html
@@ -29,9 +29,9 @@
                 <td>{{ category.id }}</td>
                 <td>{{ category.name }}</td>
                 <td>
-                    <a href="{{ url_for('edit_call_category', id=category.id) }}"
+                    <a href="{{ url_for('categories_bp.edit_call_category', id=category.id) }}"
                         class="btn btn-primary btn-sm">Редактировать</a>
-                    <a href="{{ url_for('delete_call_category_confirmation', id=category.id) }}"
+                    <a href="{{ url_for('categories_bp.delete_call_category_confirmation', id=category.id) }}"
                         class="btn btn-danger">Удалить</a>
                 </td>
             </tr>
