+++ website/templates/call_categories/delete_confirmation.html
@@ -10,13 +10,13 @@
     <p>Вы уверены, что хотите удалить категорию "{{ category.name }}"?</p>
     
     <!-- Форма для отправки данных на сервер -->
-    <form action="{{ url_for('delete_call_category', id=category.id) }}" method="POST">
+    <form action="{{ url_for('categories_bp.delete_call_category', id=category.id) }}" method="POST">
      
       <!-- Кнопка для подтверждения удаления -->
       <button type="submit" class="btn btn-danger">Удалить</button>
       
       <!-- Ссылка для возврата назад -->
-      <a href="{{ url_for('call_categories') }}" class="btn btn-secondary">Отмена</a>
+      <a href="{{ url_for('categories_bp.call_categories') }}" class="btn btn-secondary">Отмена</a>
     </form>
   </div>
 {% endblock %}
