+++ website/templates/settings/settings.html
@@ -10,8 +10,8 @@
         <label for="greeting_file" class="form-label">Файл голосового приветствия:</label>
         <input type="file" class="form-control" id="greeting_file" name="greeting_file">
         {% if settings.greeting_file %}
-        <a href="{{ url_for('get_greeting_file') }}" class="btn btn-primary mt-2">Скачать файл</a>
-        <form action="{{ url_for('delete_greeting_file') }}" method="POST" class="d-inline">
+        <a href="{{ url_for('settings_bp.get_greeting_file') }}" class="btn btn-primary mt-2">Скачать файл</a>
+        <form action="{{ url_for('settings_bpdelete_greeting_file') }}" method="POST" class="d-inline">
             <button type="submit" class="btn btn-danger mt-2">Удалить файл</button>
         </form>
         {% endif %}
