+++ website/templates/users/user_delete_confirm.html
@@ -9,6 +9,6 @@
 
 <form method="POST">
     <button type="submit" class="btn btn-danger">Да, удалить</button>
-    <a href="{{ url_for('show_users') }}" class="btn btn-secondary">Нет, вернуться</a>
+    <a href="{{ url_for('users_bp.show_users') }}" class="btn btn-secondary">Нет, вернуться</a>
 </form>
 {% endblock %}
