+++ website/templates/users/user_edit.html
@@ -65,7 +65,7 @@
     <div class="form-group row">
         <div class="offset-sm-2 col-sm-10">
             <button type="submit" class="btn btn-primary mr-2">{% if user %}Обновить{% else %}Создать{% endif %}</button>
-            <a href="{{ url_for('show_users') }}" class="btn btn-secondary mr-2">Отменить</a>
+            <a href="{{ url_for('users_bp.show_users') }}" class="btn btn-secondary mr-2">Отменить</a>
         </div>
     </div>
 </form>
