+++ website/templates/users/user_upload_file.html
@@ -19,7 +19,7 @@
         {% endif %}
     {% endwith %}
 
-    <form method="POST" action="{{ url_for('users_upload_csv') }}" enctype="multipart/form-data">
+    <form method="POST" action="{{ url_for('users_bp.users_upload_csv') }}" enctype="multipart/form-data">
         <div class="row gx-3 gy-2 align-items-center">
             <div class="col-auto">
                 <label for="csvFile" class="visually-hidden">Выберите CSV-файл:</label>
