+++ website/templates/users/users.html
@@ -22,15 +22,15 @@
     </form>
     <div class="d-flex">
         <div class="col">
-            <a href="{{ url_for('create_user') }}" class="btn btn-success mr-2">Добавить нового пользователя</a>
+            <a href="{{ url_for('users_bp.create_user') }}" class="btn btn-success mr-2">Добавить нового пользователя</a>
         </div>
         <div>&nbsp</div>
         <div class="col-*">
-            <a href="{{ url_for('users_upload_csv') }}" class="btn btn-info">Загрузить CSV-файл</a>
+            <a href="{{ url_for('users_bp.users_upload_csv') }}" class="btn btn-info">Загрузить CSV-файл</a>
         </div>        
         <div>&nbsp</div>
         <div class="col-*">
-            <a href="{{ url_for('users_export_csv') }}" class="btn btn-info">Выгрузить CSV-файл</a>
+            <a href="{{ url_for('users_bp.users_export_csv') }}" class="btn btn-info">Выгрузить CSV-файл</a>
         </div>        
     </div>
 </div>
@@ -53,8 +53,8 @@
             <td>{{ user.fio }}</td>
             <td>{{ user.phone }}</td>
             <td>
-                <a href="{{ url_for('edit_user', id=user.id) }}" class="btn btn-primary btn-sm">Редактировать</a>
-                <a href="{{ url_for('delete_user', id=user.id) }}" class="btn btn-danger">Удалить</a>
+                <a href="{{ url_for('users_bp.edit_user', id=user.id) }}" class="btn btn-primary btn-sm">Редактировать</a>
+                <a href="{{ url_for('users_bp.delete_user', id=user.id) }}" class="btn btn-danger">Удалить</a>
             </td>
         </tr>
         {% endfor %}
